requirements.txt
asgiref==3.5.2
boto3==1.23.1
botocore==1.26.2
click==8.1.3
colorama==0.4.4
Django==4.0.5
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
pathlib==1.0.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
s3transfer==0.5.2
six==1.16.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
